+++ packages/resilience/src/resilience.spec.ts
@@ -1,6 +1,8 @@
-import axios from 'axios';
+// eslint-disable-next-line import/named
+import axios, { AxiosRequestConfig, AxiosResponse } from 'axios';
 import nock from 'nock';
-import { executeWithMiddleware } from './middleware';
+import { circuitBreakerHttp, circuitBreakers } from './circuit-breaker';
+import { executeWithMiddleware, HttpMiddlewareContext } from './middleware';
 import { retry } from './retry';
 import { timeout } from './timeout';
 
@@ -12,6 +14,7 @@ describe('combined resilience features', () => {
     FORBIDDEN: 403,
     SERVICE_UNAVAILABLE: 504
   };
+  const host = 'http://example.com';
 
   it('needs to retry with delay below timeout', async () => {
     nock('https://example.com', {})
@@ -61,4 +64,46 @@ describe('combined resilience features', () => {
 
     expect(response.status).toBe(HTTP_STATUS.OK);
   }, 10000);
+  it('circuit breaker works together with a timeout', async () => {
+    const delay = 100;
+    nock(host, {})
+      .persist()
+      .get(/with-delay/)
+      .delay(delay)
+      .reply(200);
+
+    const requestConfig: AxiosRequestConfig = {
+      method: 'get',
+      baseURL: host,
+      url: 'with-delay'
+    };
+    const context: HttpMiddlewareContext = {
+      requestConfig,
+      uri: host,
+      tenantId: 'myTestTenant'
+    };
+    const request = () => axios.request(requestConfig);
+    const keepCalling = true;
+    while (keepCalling) {
+      await expect(
+        executeWithMiddleware<AxiosResponse, HttpMiddlewareContext>(
+          [timeout(delay * 0.5), circuitBreakerHttp()],
+          context,
+          request
+        )
+      ).rejects.toThrow();
+
+      const breaker = circuitBreakers[`${host}::myTestTenant`];
+      if (breaker.opened) {
+        break;
+      }
+    }
+    await expect(
+      executeWithMiddleware<AxiosResponse, HttpMiddlewareContext>(
+        [circuitBreakerHttp()],
+        context,
+        request
+      )
+    ).rejects.toThrow('Breaker is open');
+  });
 });
